+++ src/core/server/athena/main.ts
@@ -24,10 +24,12 @@ export const DEFAULT_CONFIG = {
     RESPAWN_HEALTH: 200, // Health is 100 - 200. 99 is dead.
     RESPAWN_ARMOUR: 0, // Lose weapons when you respawn
     // Command Properties
+    COMMAND_OOC_DISTANCE: 35,
     COMMAND_ME_DISTANCE: 35,
     COMMAND_DO_DISTANCE: 35,
     COMMAND_LOW_DISTANCE: 10,
     COMMAND_WHISPER_DISTANCE: 5,
+    CHAT_ROLEPLAY_OOC_COLOR: `{656565}`, // OOC Color
     CHAT_ROLEPLAY_COLOR: `{C39ADD}`, // Roleplay Color
     CHAT_ROLEPLAY_LOW_COLOR: `{d3d3d3}`, // Low Color
     CHAT_ROLEPLAY_WHISPER_COLOR: `{e6e6ce}`, // Whisper Color