+++ src/core/server/commands/commands.ts
@@ -1,3 +1,4 @@
+import './adminChat';
 import './getVehicle';
 import './noclip';
 import './revive';